Added EditboxChords
